requirements.txt
ultralytics
opencv-python
pyzbar
numpy
requests
python-telegram-bot